+++ packages/docs/src/components/DatePicker/sections.tsx
@@ -5,7 +5,6 @@ import {
   PickersUtilProvider,
 } from '@naturacosmeticos/natds-web';
 import DateFnsUtils from '@date-io/date-fns';
-import { makeStyles } from '@material-ui/core/styles';
 
 interface IDatePickerViewsProps {
   variant?: string;
@@ -17,96 +16,119 @@ interface IDatePickerViewsProps {
   animateYearScrolling?: boolean;
   autoOk?: boolean;
   disabled?: boolean;
+  required?: boolean;
   clearable?: boolean;
   views?: ['year' | 'date' | 'month'];
+  value?: Date | string | null;
+  state?: 'error' | 'success';
+  placeholder?: string;
+  helpText?: string;
 }
 
-const DatePickerWrapper = (props: IDatePickerViewsProps) => {
-  const {
-    variant,
-    disableToolbar,
-    format,
-    openTo,
-    disableFuture,
-    disablePast,
-    clearable,
-    autoOk,
-    views,
-  } = props;
-
-  const useStyles = makeStyles(theme => ({
-    wrapper: {
-      display: 'flex',
-    },
-  }));
-
-  const classes = useStyles();
+const DatePickerWrapper = ({ value = '', ...props }: IDatePickerViewsProps) => {
+  const [date, setDate] = React.useState(value);
 
   const textField = (textFieldProps: any) => (
     <TextField {...textFieldProps} id="random-prop-id" type="text" />
   );
 
   return (
-    <div className={classes.wrapper}>
-      <PickersUtilProvider utils={DateFnsUtils}>
-        <DatePicker
-          TextFieldComponent={textField}
-          variant={variant}
-          disableToolbar={disableToolbar}
-          format={format}
-          openTo={openTo}
-          disableFuture={disableFuture}
-          disablePast={disablePast}
-          clearable={clearable}
-          autoOk={autoOk}
-          views={views}
-        />
-      </PickersUtilProvider>
-    </div>
+    <PickersUtilProvider utils={DateFnsUtils}>
+      <DatePicker
+        label="Label"
+        helpText="Assistive text"
+        format="dd/MM/yyyy"
+        TextFieldComponent={textField}
+        value={date}
+        onChange={setDate}
+        {...props}
+      />
+    </PickersUtilProvider>
   );
 };
 
 export const variants = [
   {
-    title: 'Date Format',
-    component: <DatePickerWrapper variant="dialog" format="dd/MM/yyyy" />,
+    title: 'Dialog: Use for mobile web interfaces.',
+    component: <DatePickerWrapper variant="dialog" />,
   },
   {
-    title: 'Variant - Static',
-    component: <DatePickerWrapper variant="static" />,
+    title: 'Inline: Use for desktop web interfaces.',
+    component: <DatePickerWrapper variant="inline" autoOk />,
   },
+];
+
+export const types = [
   {
-    title: 'Disable Toolbar',
-    component: <DatePickerWrapper variant="static" disableToolbar={true} />,
+    title: 'Outlined',
+    component: <DatePickerWrapper />,
   },
+];
+
+export const states = [
   {
-    title: 'Auto OK',
-    component: <DatePickerWrapper variant="static" autoOk={true} />,
-  },
-  {
-    title: 'Open To',
-    component: <DatePickerWrapper variant="static" openTo="year" />,
-  },
-  {
-    title: 'Disable Future',
-    component: <DatePickerWrapper variant="static" disableFuture={true} />,
-  },
-  {
-    title: 'Disable Past',
-    component: <DatePickerWrapper variant="static" disablePast={true} />,
-  },
-  {
-    title: 'Button Clear',
+    title: 'Empty',
     component: (
-      <DatePickerWrapper
-        variant="dialog"
-        format="dd/MM/yyyy"
-        clearable={true}
-      />
+      <DatePickerWrapper value={null} placeholder="Pick a date" clearable />
     ),
   },
   {
-    title: 'Views',
-    component: <DatePickerWrapper variant="static" views={['year']} />,
+    title: 'Filled',
+    component: <DatePickerWrapper />,
+  },
+  {
+    title: 'Success',
+    component: <DatePickerWrapper state="success" />,
+  },
+  {
+    title: 'Error',
+    component: <DatePickerWrapper state="error" />,
+  },
+];
+
+export const attributes = [
+  {
+    title: 'Auto OK',
+    component: <DatePickerWrapper autoOk />,
+  },
+  {
+    title: 'Button Clear',
+    component: <DatePickerWrapper placeholder="Pick a date" clearable />,
+  },
+  {
+    title: 'Format - MM/yy',
+    component: <DatePickerWrapper format="MM/yy" />,
+  },
+  {
+    title: 'Help Text - Not used',
+    component: <DatePickerWrapper helpText="" />,
+  },
+  {
+    title: 'Disabled',
+    component: <DatePickerWrapper disabled />,
+  },
+  {
+    title: 'Required',
+    component: <DatePickerWrapper required />,
+  },
+  {
+    title: 'Disable Toolbar',
+    component: <DatePickerWrapper variant="static" disableToolbar />,
+  },
+  {
+    title: 'Disable Future',
+    component: <DatePickerWrapper variant="static" disableFuture />,
+  },
+  {
+    title: 'Disable Past',
+    component: <DatePickerWrapper variant="static" disablePast />,
+  },
+  {
+    title: 'Open To - Year',
+    component: <DatePickerWrapper variant="static" openTo="year" />,
+  },
+  {
+    title: 'Views - Month',
+    component: <DatePickerWrapper variant="static" views={['month']} />,
   },
 ];